More GCC compile warnings (#37)
* ci: clang format on push

* feat: more compile warnings

* fix: add unknown clang warning flags only for gcc

* chore: remove -Winline

* fix: size_t instead of unsigned int

* fix: explicit casting of SpellLevel enum to unsigned int

+++ src/models/character.cpp
@@ -38,7 +38,7 @@ const std::unordered_map<std::string, int> dnd::Character::getInitialAttributeVa
         {"MAXHP", 0},
         {"ARMOR_ON", false},
     };
-    for (int i = 0; i < 6; ++i) {
+    for (size_t i = 0; i < 6; ++i) {
         character_initial_values.emplace(ability_strings_inorder[i], base_ability_scores[i] * 100);
         character_initial_values.emplace(ability_strings_inorder[i] + "MAX", 2000);
     }
